+++ templates/app/Models/UserPermission.php
@@ -3,17 +3,21 @@
  * This file is cloned / force-published from the "laravel-auth0-pattern" composer package.
  *    WARNING: Local modifications will be overwritten when the package is updated.
  *             See https://github.com/faithfm/laravel-auth0-pattern for more details.
+ * 
+ *    NOTE:    This file supercedes the UserPermission.php file published by the "faithfm/laravel-simple-permissions" package.
+ *             It is identical to the upstream file, except for the inclusion of support for the owen-it/laravel-auditing package.
  */
 
 namespace App\Models;
 
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
-use OwenIt\Auditing\Contracts\Auditable;
+use OwenIt\Auditing\Auditable as AuditableTrait;
+use OwenIt\Auditing\Contracts\Auditable as AuditableContract;
 
-class UserPermission extends Model implements Auditable
+class UserPermission extends Model implements AuditableContract
 {
-    use \OwenIt\Auditing\Auditable;
+    use AuditableTrait;
     use HasFactory;
 
     /**