+++ templates/config/auth.php
@@ -13,14 +13,15 @@ return [
     | Authentication Defaults
     |--------------------------------------------------------------------------
     |
-    | This option defines the default authentication "guard" and password
-    | reset "broker" for your application. 
+    | This option controls the default authentication "guard" and password
+    | reset options for your application. You may change these defaults
+    | as required, but they're a perfect start for most applications.
     |
     */
 
     'defaults' => [
-        'guard' => env('AUTH_GUARD', 'ffm-session-guard'),
-        'passwords' => env('AUTH_PASSWORD_BROKER', 'users'),
+        'guard' => 'ffm-session-guard',
+        'passwords' => 'users',
     ],
 
     /*
@@ -29,41 +30,38 @@ return [
     |--------------------------------------------------------------------------
     |
     | Next, you may define every authentication guard for your application.
+    | Of course, a great default configuration has been defined for you
+    | here which uses session storage and the Eloquent user provider.
     |
-    | All authentication guards have a user provider, which defines how the
+    | All authentication drivers have a user provider. This defines how the
     | users are actually retrieved out of your database or other storage
-    | system used by the application.
+    | mechanisms used by this application to persist your user's data.
+    |
+    | Our faithfm/laravel-auth0-pattern package renames the default Laravel guards 
+    | for clarity, searchability, and disambiguation with route middleware naming:
+    |   * 'web' guard is renamed to 'ffm-session-guard'
+    |   * 'api' guard is renamed to 'ffm-token-guard'
     |
     */
 
     'guards' => [
-        // Faith FM's session guard (Auth0 session driver + an Eloquent-linked user-provider)
+        // Renamed, but identical to Laravel's default 'web' session guard
+        // (The actual login is performed by CallbackController in laravel-simple-auth0)
         'ffm-session-guard' => [
-          'driver' => 'auth0.authenticator.patched',    // patched to fix the issue with accessToken vs idToken when AUTH0_AUDIENCE is blank (ie: using the default https://faithfm.au.auth0.com/userinfo endpoint instead of an API endpoint)
-          'provider' => 'ffm-auth0-user-provider',      // based on our custom User Repository
-          'configuration' => 'web',                     // not documented well, but this setting points Auth0's configurator to use the 'guards.web' section defined in 'config/auth0.php' - see config('auth0.guards.' . $this->guardConfigurationKey) in InstanceEntityAbstract.php
+            'driver' => 'session',
+            'provider' => 'users',
         ],
 
-        // // Auth0's default session guard (does not implement an Eloquent-linked user provider) - NOT NORMALLY USED IN OUR APPS
-        // 'auth0-session' => [
-        //   'driver' => 'auth0.authenticator',
-        //   'provider' => 'auth0-provider',
-        //   'configuration' => 'web',
-        // ],
-
-        // Faith FM's token-based guard (same as Larave's default token-based guard - renamed from 'api' to 'ffm-token-guard' to avoid confusion + increase searchability
+        // Renamed, but identical to Laravel's default 'api' token guard 
+        // (Still supported although config not included by default in new projects since Laravel 7)
+        // (Database migrations laravel-simple-auth-tokens package and User model changes documented in the )
         'ffm-token-guard' => [
             'driver' => 'token',
-            'provider' => 'eloquent_users',
-            'hash' => false,
+            'provider' => 'users',
+            'input_key' => 'api_token',       // Default value - shown here for clarity
+            'storage_key' => 'api_token',     // Default value - shown here for clarity
+            'hash' => false,                  // Default value - shown here for clarity
         ],
-
-        // // Auth0's default token guard - NOT NORMALLY USED IN OUR APPS
-        // 'auth0-api' => [
-        //     'driver' => 'auth0.authorizer',
-        //     'configuration' => 'api',
-        //     'provider' => 'auth0-provider',
-        // ],
     ],
 
     /*
@@ -71,29 +69,23 @@ return [
     | User Providers
     |--------------------------------------------------------------------------
     |
-    | All authentication guards have a user provider, which defines how the
+    | All authentication drivers have a user provider. This defines how the
     | users are actually retrieved out of your database or other storage
-    | system used by the application. 
+    | mechanisms used by this application to persist your user's data.
+    |
+    | If you have multiple user tables or models you may configure multiple
+    | sources which represent each model / table. These sources may then
+    | be assigned to any extra authentication guards you have defined.
+    |
+    | Supported: "database", "eloquent"
     |
     */
 
     'providers' => [
-        // Faith FM's user-provider (including Auth0 information + Eloquent User model information, by means of our Auth0PatternUserRepository)
-        'ffm-auth0-user-provider' => [
-            'driver' => 'auth0.provider',
-            'repository' => FaithFM\Auth0Pattern\Auth0PatternUserRepository::class,
-        ],
-
-        // // Auth0's default user-provider (when Auth0 information only is required - ie: when not using an Eloquent User model) - NOT NORMALLY USED IN OUR APPS
-        // 'auth0-provider' => [
-        //   'driver' => 'auth0.provider',
-        //   'repository' => 'auth0.repository',
-        // ],
-
-        // Laravel's default user-provider - used by our 'ffm-token-guard' guard
-        'eloquent_users' => [
+        // Laravel's default user-provider configuration
+        'users' => [
             'driver' => 'eloquent',
-            'model' => env('AUTH_MODEL', App\Models\User::class),
+            'model' => App\Models\User::class,
         ],
     ],
 
@@ -102,9 +94,9 @@ return [
     | Resetting Passwords
     |--------------------------------------------------------------------------
     |
-    | These configuration options specify the behavior of Laravel's password
-    | reset functionality, including the table utilized for token storage
-    | and the user provider that is invoked to actually retrieve users.
+    | You may specify multiple password reset configurations if you have more
+    | than one user table or model in the application and you want to have
+    | separate password reset settings based on the specific user types.
     |
     | The expiry time is the number of minutes that each reset token will be
     | considered valid. This security feature keeps tokens short-lived so
@@ -114,7 +106,7 @@ return [
     | generating more password reset tokens. This prevents the user from
     | quickly generating a very large amount of password reset tokens.
     |
-    | CURRENTLY UNUSED in Faith FM pattern (we use Auth0's password reset functionality)
+    | CURRENTLY UNUSED in Faith FM pattern (Auth0 is responsible for password reset functionality)
     |
     */
 
@@ -122,7 +114,7 @@ return [
     'passwords' => [
         'users' => [
             'provider' => 'users',
-            'table' => env('AUTH_PASSWORD_RESET_TOKEN_TABLE', 'password_reset_tokens'),
+            'table' => 'password_reset_tokens',
             'expire' => 60,
             'throttle' => 60,
         ],
@@ -134,7 +126,7 @@ return [
     |--------------------------------------------------------------------------
     |
     | Here you may define the amount of seconds before a password confirmation
-    | window expires and users are asked to re-enter their password via the
+    | times out and the user is prompted to re-enter their password via the
     | confirmation screen. By default, the timeout lasts for three hours.
     |
     | CURRENTLY UNUSED in Faith FM pattern (we use Auth0's password reset functionality)
@@ -142,7 +134,7 @@ return [
     */
 
     // Laravel's default password confirmation timeout (NOT IN USE)
-    'password_timeout' => env('AUTH_PASSWORD_TIMEOUT', 10800),
+    'password_timeout' => 10800,
 
     /*
     |--------------------------------------------------------------------------
@@ -152,8 +144,8 @@ return [
     | The list of authorization permissions recognised by the application 
     | (as applied to each user in the 'user_permissions' table).
     |
-    | Faith FM laravel-auth0-pattern automatically creates Gates for all permissions 
-    | defined here.  (See: Auth0PatternServiceProvider.php)
+    | Faith FM laravel-simple-permissions automatically creates Gates for all permissions 
+    | defined here.  (See: SimplePermissionsServiceProvider.php)
     |
     | Ie: permissions applied in the 'user_permissions' table must be defined here
     | before they will be available for Authorization in the application.